# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "contact": {
            "name": "Support Team",
            "email": "support@my-api.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/talla": {
            "get": {
                "tags": [
                    "Talla"
                ],
                "summary": "Obtener lista de tallas",
                "description": "Retorna una lista de tallas",
                "operationId": "d8fafae24f203fe0967ebbd4a3ff88bc",
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "Descripcion": {
                                                "description": "Descripción de la talla",
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Talla"
                ],
                "summary": "Crear una nueva talla",
                "description": "Crea una nueva talla con los datos enviados",
                "operationId": "bb8fa6710b06b02a8e059c935f477b67",
                "requestBody": {
                    "description": "Información de la talla para crear un nuevo item",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "Descripcion": {
                                        "description": "Descripción de la talla",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Talla creada correctamente"
                    },
                    "default": {
                        "description": "Ha ocurrido un error"
                    }
                }
            }
        },
        "/api/talla/{id}": {
            "get": {
                "tags": [
                    "Talla"
                ],
                "summary": "Obtener una talla",
                "description": "Retorna una talla por su ID",
                "operationId": "b39c17989ff11a6347ede401dddfc963",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la talla",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Descripcion": {
                                            "description": "Descripción de la talla",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Talla"
                ],
                "summary": "Actualizar una talla",
                "description": "Actualiza una talla con los datos enviados",
                "operationId": "9cdf103dbe4e73af4a0ef4afd00c64ce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la talla a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Información de la talla para actualizar un item",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "Descripcion": {
                                        "description": "Descripción de la talla",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Talla actualizada correctamente"
                    },
                    "default": {
                        "description": "Ha ocurrido un error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Talla"
                ],
                "summary": "Eliminar una talla",
                "description": "Elimina una talla por su ID",
                "operationId": "70b6a8e3ab05ce3fe0ce58eed39d614d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la talla a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Talla eliminada correctamente"
                    },
                    "default": {
                        "description": "Ha ocurrido un error"
                    }
                }
            }
        },
...
    },
...
}